cfg-setters: code-style packaging.
